Completed Day 13

+++ day-13/src/main/java/com/someware/aoc2024/day13/Vec.java
@@ -1,4 +1,4 @@
 package com.someware.aoc2024.day13;
 
-public record Vec(int x, int y) {
+public record Vec(long x, long y) {
 }
